template OptData for better data management
