+++ modules/optimization/include/optimizationreporters/OptimizationReporter.h
@@ -9,35 +9,23 @@
 
 #pragma once
 
-#include "OptimizationDataHelper.h"
+#include "OptimizationData.h"
 #include "OptimizationReporterBase.h"
 
 /**
  * Computes gradient and contains reporters for communicating between optimizeSolve and subapps
  */
-class OptimizationReporter : public OptimizationReporterBase
+class OptimizationReporter : public OptimizationDataTempl<OptimizationReporterBase>
 {
 
 public:
   static InputParameters validParams();
   OptimizationReporter(const InputParameters & parameters);
 
-  void execute() override;
-
   virtual Real computeObjective() override;
   virtual void setMisfitToSimulatedValues() override;
 
 protected:
-  /// Data helper for generating objective value.
-  OptimizationDataHelper _opt_data;
-
-  /// measurement values
-  std::vector<Real> & _measurement_values;
-  /// simulated values at measurement xyzt
-  std::vector<Real> & _simulation_values;
-  /// difference between simulation and measurement values at measurement xyzt
-  std::vector<Real> & _misfit_values;
-
 private:
   void setICsandBounds();
   virtual void setSimulationValuesForTesting(std::vector<Real> & data) override;
